Improving certain aspects of numpy types serialization/deserialization
Prior to this commit, numpy was serialized with np.save
without support for pickle fallback. This was a compromise
between efficiency and flexibility.

With this changes, numpy.ndarray have a straightforward
serialization (which loses a little bit of flexibility) while
np.save is used for other numpy types (gaining flexibility).